Moved BIBI Configuration to CLE repository

+++ hbp_nrp_backend/hbp_nrp_backend/exd_config/tests/test_experiment_configuration_script.py
@@ -32,23 +32,14 @@ class TestExperimentConfigurationScript(unittest.TestCase):
         directory = os.path.split(__file__)[0]
         experiment = os.path.join(directory, 'ExDXMLExample.xml')
         generated_bibi = os.path.join(directory, 'generated_bibi.py')
-        expected_generated_bibi = os.path.join(directory, 'expected_bibi.py')
 
         # Remove the generated file if it already exists.
-        try:
+        if os.path.isfile(generated_bibi):
             os.remove(generated_bibi)
-        except OSError:
-            pass
 
         # Generate bibi script file and compare it to an expected file.
         generate_bibi(experiment, generated_bibi)
-        self.assertTrue(os.path.exists(generated_bibi))
-        file1 = open(generated_bibi, 'r')
-        file2 = open(expected_generated_bibi, 'r')
-        diff = difflib.context_diff(file1.readlines(), file2.readlines())
-        delta = ''.join(diff)
-        self.maxDiff = None
-        self.assertMultiLineEqual(delta, "")
+        self.assertTrue(os.path.isfile(generated_bibi))
 
 if __name__ == '__main__':
     unittest.main()
